backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-settings==2.6.0
supabase==2.10.0
boto3==1.35.0
upstash-redis==1.1.0
httpx==0.27.0
python-multipart==0.0.12
Pillow==10.4.0
rembg==2.0.57
modal==0.64.0
razorpay==1.4.2
resend==2.4.0
python-jose[cryptography]==3.3.0
bleach==6.2.0
